requirements.txt
matplotlib==3.9.0
numpy==1.26.4
opencv-python==4.10.0.82
ultralytics==8.2.32
Pillow==10.3.0
